Towards #38 and #39 - Connected Frontend and Backend Form
Updated Report Model and Form so fields are automatically generated from the Report Model, as well much of the validation process (specifically fields which are strictly required no matter the circumstances). Report Form has been updated to primarily handle circumstantial validation.

View.py has been updated with the Submit_Report method to handle the report checking as well as adding any necessary information to the report data before it saves the data to the database.

urls.py has been updated to accommodate the new submit_report method and it's path

Admin.py has been updated to accommodate the updates to the Report Model and it's fields.

input_form.html has been updated to properly send data to the form via it's POST method. As well as the additon of a simple submit button that has to be visually updated. Along with a few small error fixes.