Towards #70 - Updated Method to Keep Data after Error
Updated input_form.html to allow already formatted form data to be displayed so when an error is encountered the user won't have to enter the data again.

Updated views.py to check for a POST method which means a user is submitting a form, or saving a form, which will be implemented in a later issue.