Towards #61
Updated models.py to handle reports with no information, allowing fields to be saved as null.

Updated forms.py's clean method to be the validate method. This method will handle the extra validation required for a submitted form, as well as checking that the necessary fields are apparent.

Updated views.py to handle a 'save' request, adding the Partially Completed status to the report and saving it without checking for required fields.