Closes #108 - Fixed mark_report_complete and sign_off_report view methods
Added redirects to mark_report_complete and sign_off_report views, as well as more descriptive error messages and adding the username of the account that marked the report as complete and signed off on the report.

Added additional information to be displayed on the read_only_report.html template. This includes report status, submission date (Automatically generated when first saved/submitted), and the username of those attached to the report.