Closes #109 - Update Name of Writer Fields and More
Updated Report Model to include fields for First and Last Name of the Report Writer as well as their position, and removed the name_of_writer field which the new fields replace.

Update Report Form to include checks for the new fields for a validated form.

Update all input forms with input fields for all of the fields, as well as updated the dashboard to include a formatted version of the writer's name. Also updated the date of incident line to check if a date of incident has been entered and if not the dashboard uses the automatically included submission date.