Closes #110 - Updated Incident Location Field
Updated Report Model so that the incident location field has a list of choices which the user can choose. As well a CharField to describe a location other than those listed was added

Updated Report Form validation method to check for the new incident location field and the other description if necessary.

Updated Dashboards to correctly display the location information as well as the input forms to properly input the value of the incident location.

Update Report Form to include checks for the new fields for a validated form.

Update all input forms with input fields for all of the fields, as well as updated the dashboard to include a formatted version of the writer's name. Also updated the date of incident line to check if a date of incident has been entered and if not the dashboard uses the automatically included submission date.