Closes #116 - Implement Unauthenticated User Navigation Tracking
Added session data to track which page a user was using when their authentication ran out and upon successful login they will be navigated back to the previous page.

If the previous page was one of the editable input forms the information will also be saved and displayed for them to continue editing.